Add staff indices to notes & staves, and add note array to staves.
Notes now contain a "staff index", with C0 being at index 0, D0 at
index 1, and so on.

The step from a staff line to an adjoining staff space is a "staff
step".  The distance from one staff line to the next line is therefore
2 staff steps.

The difference between the staff indices of two notes, is equal to the
distance, in staff steps, between those notes when rendered on a
staff.

Each staff (treble, alto, bass) is aware of the staff index of its
bottom staff line. This will be used later on for rendering.

Also, a staff now has an array of notes. This will allow simple
individual-note (non-chord) rendering for now. Chord support will only
be added when needed.